+++ FirebaseVertexAI/CHANGELOG.md
@@ -9,6 +9,8 @@
   is now handled by the existing `unknown` case. (#13699)
 - [changed] **Breaking Change**: The `data` case in the `Part` enum has been
   renamed to `inlineData`; no functionality changes. (#13700)
+- [changed] **Breaking Change**: The property `citationSources` of
+  `CitationMetadata` has been renamed to `citations`. (#13702)
 
 # 11.3.0
 - [added] Added `Decodable` conformance for `FunctionResponse`. (#13606)